Shop/models.py
from django.db import models
from django.contrib.auth.models import User
import datetime
import os
from PIL import Image

def getFileName (request,filename):
    now_time = datetime.datetime.now().strftime("%Y%m%d%H:%M:%S")
    new_filename = "%s%s"%(now_time,filename)
    return os.path.join('uploads/',new_filename)

class Category(models.Model):
    name = models.CharField(max_length = 150, null = False, blank=False)
    image = models.ImageField(upload_to=getFileName, null= True, blank= True)
    description = models.TextField(max_length = 500, null = False, blank = False )
    status = models.BooleanField(default = False, help_text='0-show, 1-Hidden')
    created_at = models.DateTimeField(auto_now_add = True)

    def __str__(self):
        return self.name
    

# class Product(models.Model):
#     category = models.ForeignKey(Category, on_delete = models.CASCADE)
#     name = models.CharField(max_length = 150, null = False, blank=False)
#     vendor = models.CharField(max_length = 150, null = False, blank = False)
#     product_image = models.ImageField(upload_to=getFileName, null= True, blank= True)
#     quantity = models.IntegerField( null = False, blank = False )
#     original_price = models.FloatField( null = False, blank = False )
#     selling_price = models.FloatField( null = False, blank = False )
#     description = models.TextField(max_length = 500, null = False, blank = False )
#     status = models.BooleanField(default = False, help_text='0-show, 1-Hidden')
#     trending = models.BooleanField(default= False, help_text= '0-default, 1-Trending')
#     created_at = models.DateTimeField(auto_now_add = True)

#     def __str__(self):
#         return self.name

class Product(models.Model):
    category = models.ForeignKey(Category, on_delete=models.CASCADE)
    name = models.CharField(max_length=150, null=False, blank=False)
    vendor = models.CharField(max_length=150, null=False, blank=False)
    product_image = models.ImageField(upload_to='products')
    quantity = models.IntegerField(null=False, blank=False)
    original_price = models.FloatField(null=False, blank=False)
    selling_price = models.FloatField(null=False, blank=False)
    description = models.TextField(max_length=500, null=False, blank=False)
    status = models.BooleanField(default=False, help_text='0-show, 1-Hidden')
    trending = models.BooleanField(default=False, help_text='0-default, 1-Trending')
    created_at = models.DateTimeField(auto_now_add=True)

    def __str__(self):
        return self.name

    # ---------- ADD THIS PART BELOW ----------
    def save(self, *args, **kwargs):
        super().save(*args, **kwargs)

        if self.product_image:
            image_path = self.product_image.path
            img = Image.open(image_path)

            # Your fixed size → (600, 600) OR (500, 500)
            output_size = (300, 300)

            # Resize & Save
            img = img.resize(output_size)
            img.save(image_path)


class Cart(models.Model) :
    user = models.ForeignKey(User, on_delete = models.CASCADE)
    product = models.ForeignKey(Product, on_delete = models.CASCADE)
    product_quantity = models.IntegerField(null = False, blank = False)
    created_at = models.DateTimeField(auto_now_add = True)

    @property
    def total_cost(self) :
        return self.product_quantity * self.product.selling_price
    

class Favourite(models.Model) :
    user = models.ForeignKey(User, on_delete = models.CASCADE)
    product = models.ForeignKey(Product, on_delete = models.CASCADE)
    created_at = models.DateTimeField(auto_now_add = True)
