requirements.txt
asgiref==3.10.0
certifi==2025.10.5
charset-normalizer==3.4.4
cloudinary==1.44.1
distlib==0.4.0
dj-database-url==3.0.1
Django==5.2.8
django-cloudinary-storage==0.3.0
django-jazzmin==3.0.1
filelock==3.20.0
gunicorn==23.0.0
idna==3.11
mysqlclient==2.2.7
packaging==25.0
pillow==12.0.0
pipenv==2025.0.4
platformdirs==4.5.0
requests==2.32.5
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
virtualenv==20.35.4
whitenoise==6.11.0